feat(order): add amount field in create order response data

+++ internal/usecase/order/create_order.go
@@ -6,7 +6,7 @@ import (
 	paymentModel "synapsis-challange/internal/model/payment"
 )
 
-func (u orderUsecase) CreateOrder(ctx context.Context, userId int32) (int32, string, int32, error) {
+func (u orderUsecase) CreateOrder(ctx context.Context, userId int32) (int32, string, float64, error) {
 	items, err := u.cartItemRepository.ListCartItemsByUserID(ctx, userId)
 	if err != nil {
 		return 0, "", 0, err
