+++ internal/usecase/order/usecase.go
@@ -5,5 +5,5 @@ import (
 )
 
 type Usecase interface {
-	CreateOrder(ctx context.Context, userId int32) (int32, string, int32, error)
+	CreateOrder(ctx context.Context, userId int32) (int32, string, float64, error)
 }
